Added some styles to header
