+++ src/css/quasar.variables.scss
@@ -18,7 +18,7 @@ $accent: #7e8b77;
 
 $dark: #1d1d1d;
 
-$positive: #21ba45;
+$positive: #78981e;
 $negative: #c10015;
 $info: #31ccec;
 $warning: #f2c037;
