+++ src/pages/Home/components/HomeHeader/styles.scss
@@ -1,0 +1,49 @@
+.Header {
+  background-color: $secondary;
+  border-radius: 5px;
+  padding: 1.95rem 1rem 1.3rem 1.95rem;
+  display: flex;
+  justify-content: space-between;
+
+  &__greetings {
+    font-size: 14px;
+  }
+
+  &__names {
+    font-size: 1.125rem;
+    font-weight: bolder;
+  }
+
+  &__text {
+    font-size: 0.75rem;
+
+    p {
+      margin-bottom: 0.625rem;
+    }
+  }
+
+  &__membership {
+    &--active {
+      color: $positive;
+    }
+
+    &--negative {
+      color: $negative;
+    }
+  }
+
+  &__avatar {
+    width: 3.5rem;
+    height: 3.5rem;
+  }
+
+  &__editIcon {
+    position: absolute;
+    bottom: -5px;
+    right: -5px;
+    background-color: white;
+    border-radius: 50%;
+    padding: 2px;
+    box-shadow: rgba(0, 0, 0, 0.16) 0px 1px 4px;
+  }
+}
