+++ src/pages/Home/components/index.js
@@ -1,4 +1,4 @@
-import HomeHeader from './HomeHeader.vue';
+import HomeHeader from './HomeHeader/HomeHeader.vue';
 import HomeProviders from './HomeProviders.vue';
 import HomeEvents from './HomeEvents.vue';
